backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
pydantic
tavily-python
python-dotenv
google-generativeai
